@@ -9,15 +9,17 @@
   },
   "dependencies": {
     "babel-polyfill": "^6.23.0",
+    "create-react-class": "^15.5.1",
     "date-fns": "^1.28.2",
     "font-awesome": "^4.7.0",
     "lodash": "^4.17.4",
     "normalize.css": "^6.0.0",
+    "prop-types": "^15.5.4",
     "radium": "^0.18.2",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "react": "^15.5.3",
+    "react-dom": "^15.5.3",
     "react-hot-loader": "^3.0.0-beta.6",
-    "react-onclickoutside": "^5.10.0",
+    "react-onclickoutside": "^5.11.0",
     "react-redux": "^5.0.3",
     "recharts": "^0.22.0",
     "redux": "^3.6.0",
@@ -25,12 +27,12 @@
     "whatwg-fetch": "^2.0.3"
   },
   "devDependencies": {
-    "babel-core": "^6.24.0",
+    "babel-core": "^6.24.1",
     "babel-eslint": "^7.2.1",
     "babel-loader": "^6.4.1",
-    "babel-preset-env": "^1.3.2",
-    "babel-preset-react": "^6.23.0",
-    "babel-preset-stage-0": "^6.22.0",
+    "babel-preset-env": "^1.3.3",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
     "css-loader": "^0.28.0",
     "eslint": "^3.19.0",
     "eslint-config-airbnb": "^14.1.0",
