@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
 import DonutChart from './DonutChart';
@@ -7,9 +8,7 @@ import { fetchTransactions } from '../reducers/transactions';
 
 const styles = {
   container: {
-    height: '100vh',
     maxWidth: 800,
-    overflow: 'hidden',
     margin: '0 auto',
   },
 };
@@ -30,7 +29,7 @@ class App extends React.Component {
 }
 
 App.propTypes = {
-  fetchTransactions: React.PropTypes.func.isRequired,
+  fetchTransactions: PropTypes.func.isRequired,
 };
 
 export default connect(null, { fetchTransactions })(App);
